+++ changelog.md
@@ -5,3 +5,8 @@
 # V2.0.2
 - [CMakeLists.txt] BUGFIXED.
 - 添加一些注释，优化代码格式等。
+
+# V2.0.3
+- [simulator.cpp/hpp] RENAME: duration -> endtime.
+- [packmodules.cpp] BUGFIXED.
+- 添加一些日志打印。